Allow for A/B differentiation on Lattice ECP5 Evaluation Board
The FT2232H on the ECP5 evaluation board has two interfaces; A is connected
to JTAG and B to UART/I2C (see figure A.3 in the board's user guide). The
FTDI driver apparently offers up these interfaces independently by
appending " A" and " B" to the USB description ("Lattice ECP5 Evaluation
Board"). Add detection of the "A" port.